refactor(browser): split tool routing lifecycle and screenshots
Move existing browser tool behavior into internal concept modules and remove the grandfathered max-lines suppression and baseline entry.
